+++ internal/sdkv2/morpheus/resources/task/morpheus_task_python_script_resource_git_test.go
@@ -3,47 +3,31 @@
 package task_test
 
 import (
-	"context"
-	"os"
+	"strings"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
-	"github.com/hashicorp/terraform-plugin-mux/tf5to6server"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 
 	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
-	sdkv2morpheus "github.com/HPE/terraform-provider-hpe/internal/sdkv2/morpheus"
 )
 
-func TestMain(m *testing.M) {
-	code := m.Run()
-
-	testhelpers.WriteMergedResults()
-
-	os.Exit(code)
-}
-
-func newProviderWithError() (tfprotov6.ProviderServer, error) {
-	return tf5to6server.UpgradeServer(context.Background(), sdkv2morpheus.Provider().GRPCProvider)
-}
-
-var testAccProtoV6ProviderFactories = map[string]func() (
-	tfprotov6.ProviderServer, error,
-){
-	"hpe": newProviderWithError,
-}
-
-func RenderTaskPythonScriptConfig(t *testing.T, overrides map[string]string) (string, error) {
+func RenderMorpheusTaskPythonScriptGitConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
 	t.Helper()
 
-	name := acctest.RandomWithPrefix(t.Name())
 	defaults := map[string]string{
 		"Name":               name,
-		"Code":               name,
+		"Code":               strings.ToLower(name),
 		"Labels":             "[\"demo\", \"terraform\"]",
-		"SourceType":         "local",
-		"ScriptContent":      "print('morpheus')\\nprint('python')",
+		"SourceType":         "repository",
+		"ResultType":         "json",
+		"ScriptPath":         "example.py",
+		"VersionRef":         "master",
+		"RepositoryId":       "0",
 		"CommandArguments":   "example",
 		"AdditionalPackages": "pyyaml",
 		"PythonBinary":       "/usr/bin/python3",
@@ -59,12 +43,15 @@ func RenderTaskPythonScriptConfig(t *testing.T, overrides map[string]string) (st
 
 	return testhelpers.RenderExample(
 		t,
-		"task_python_script_resource.tf.tmpl",
+		"morpheus_task_python_script_resource_git.tf.tmpl",
 		"Name", defaults["Name"],
 		"Code", defaults["Code"],
 		"Labels", defaults["Labels"],
 		"SourceType", defaults["SourceType"],
-		"ScriptContent", defaults["ScriptContent"],
+		"ResultType", defaults["ResultType"],
+		"ScriptPath", defaults["ScriptPath"],
+		"VersionRef", defaults["VersionRef"],
+		"RepositoryId", defaults["RepositoryId"],
 		"CommandArguments", defaults["CommandArguments"],
 		"AdditionalPackages", defaults["AdditionalPackages"],
 		"PythonBinary", defaults["PythonBinary"],
@@ -75,7 +62,7 @@ func RenderTaskPythonScriptConfig(t *testing.T, overrides map[string]string) (st
 	)
 }
 
-func TestAccMorpheusTaskPythonScriptExampleOk(t *testing.T) {
+func TestAccMorpheusTaskPythonScriptGitExampleOk(t *testing.T) {
 	t.Parallel()
 
 	defer testhelpers.RecordResult(t)
@@ -88,10 +75,7 @@ func TestAccMorpheusTaskPythonScriptExampleOk(t *testing.T) {
 
 	name := acctest.RandomWithPrefix(t.Name())
 
-	resourceConfig, err := RenderTaskPythonScriptConfig(t, map[string]string{
-		"Name": name,
-		"Code": name,
-	})
+	resourceConfig, err := RenderMorpheusTaskPythonScriptGitConfig(t, name, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -107,7 +91,7 @@ func TestAccMorpheusTaskPythonScriptExampleOk(t *testing.T) {
 		resource.TestCheckResourceAttr(
 			resourceName,
 			"code",
-			name,
+			strings.ToLower(name),
 		),
 		resource.TestCheckResourceAttr(
 			resourceName,
@@ -122,12 +106,27 @@ func TestAccMorpheusTaskPythonScriptExampleOk(t *testing.T) {
 		resource.TestCheckResourceAttr(
 			resourceName,
 			"source_type",
-			"local",
+			"repository",
 		),
 		resource.TestCheckResourceAttr(
 			resourceName,
-			"script_content",
-			"print('morpheus')\\nprint('python')\\n",
+			"result_type",
+			"json",
+		),
+		resource.TestCheckResourceAttr(
+			resourceName,
+			"script_path",
+			"example.py",
+		),
+		resource.TestCheckResourceAttr(
+			resourceName,
+			"version_ref",
+			"master",
+		),
+		resource.TestCheckResourceAttr(
+			resourceName,
+			"repository_id",
+			"0",
 		),
 		resource.TestCheckResourceAttr(
 			resourceName,
