+++ internal/sdkv2/morpheus/resources/task/morpheus_task_python_script_resource_url.tf.tmpl
